+++ test/open_app.c
@@ -21,6 +21,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <unistd.h>
 #include <sys/types.h>
 #include <string.h>
@@ -72,7 +73,7 @@ void print_usage(char *progname)
 
 static int __launch_app_dead_handler(int pid, void *data)
 {
-	int listen_pid = (int) data;
+	int listen_pid = (intptr_t)data;
 
 	if(listen_pid == pid)
 		g_main_loop_quit(mainloop);
@@ -94,7 +95,7 @@ static gboolean run_func(void *data)
 		str = bundle_get_val(kb, "__LAUNCH_APP_MODE__");
 
 		if (str && strcmp(str, "SYNC") == 0 )
-			aul_listen_app_dead_signal(__launch_app_dead_handler, (void *)pid);
+			aul_listen_app_dead_signal(__launch_app_dead_handler, (void *)(intptr_t)pid);
 		else
 			g_main_loop_quit(mainloop);
